requirements.txt
Flask
SQLAlchemy
requests
python-dotenv
matplotlib
ping3